requirements.txt
asgiref==3.2.7
autopep8==1.5.3
braintree==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.0.7
Django==3.0.7
django-allauth==0.42.0
django-better-admin-arrayfield==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
